drop bash from scad generating dune rule
target specification is not working for me with current dune, but this
should be patched soon. Then I can use sandboxing/promotion to move the
scads into the source tree if wanted.

TODO: A rule that also exports to stls. Those are what I would really
want promoted, rather than scads, so that they can be viewed on the repo.